Implement parseFileWithExts
This file generalizes parseFile by being parametrized on a list of 
extensions that should be recognized when parsing. parseFile currently 
defaults to no extensions enabled, though note that this (currently) 
doesn't mean that no extensions will be read, since I've only just 
started scratching the surface on implementing the actual 
parametrization in the parser.