Descriptive error messages in checkEnabled(OneOf).
Rationale: I have enabled stylish-haskell in my IntelliJ and was greeted with
the existing error message. Initially I thought I had to enable some sort of
stylish-haskell plugin and started googling for a solution.
Only after trying again on a different file, I realised the error
message was complaining about a missing language pragma: the second file
was missing a pragma which I recognised the name of (the first one
I did not know). Only then did I realise that it is not stylish-haskell
that is complaining, but rather it is this code. Feel free to adjust the wording
as long as the spirit of the change is maintained: giving clear error message
as to what happened and what to do to get it to work.