Make the fields of SrcLoc and SrcSpan strict
A lot of the allocations when using HSE are caused by SrcSpans and
SrcLocs. By making the fields strict, GHC can unbox them and hence we
save a number of pointers every time a SrcLoc or SrcSpan is allocated.

There is no empirical measurments to justify this change but
a similar change was introduced in GHC and compiler allocations fell by
a non-trivial amount as a result.